requirements.txt
fastapi 
uvicorn 
python-telegram-bot 
pymongo 
dnspython
